Removed unused methods Added descriptive commenting Added type-hinting Corrected regex error in BillIdentifier Class Added error-catching in BillIdentifier Class Renamed some variables and class names for clarity
There was a debate over whether to split rss_etl into separate scripts for better readability. After much thought and prayer I concluded that, for the moment, I'll keep everything related to the rss_etl in one script, with the option to migrate and separate them later. The code is 250 lines, and everything is closely related.

In terms of error handling, I added additional error catching only for the regex in BillIdentifier, mostly because the basic regex error is confusing and nondescriptive. With everything else, I'm content to leave things as is, with intent to reevaluate later if we get a lot of noisy breaking, or people propose some useful error catching.